+++ src/Components/Atoms/ConnectBtn.tsx
@@ -1,0 +1,7 @@
+import styled from "@emotion/styled";
+
+export const ConnectJoinBtn = styled.button`
+  border: 1px solid #0a66c2;
+  background-color: transparent;
+  color: #0a66c2;
+`;
