+++ src/Components/Atoms/Heading.tsx
@@ -1,0 +1,15 @@
+import styled from "@emotion/styled";
+
+export const FormHeading = styled.h1`
+  font-size: 2.2rem;
+  margin: 0;
+  padding: 0;
+  line-height: 0.9;
+  font-weight: 500;
+  color: #000;
+  @media only screen and (min-width: 768px) {
+    font-size: 1.8rem;
+    line-height: 1.1;
+    font-weight: 600;
+  }
+`;
