+++ src/Components/Atoms/HomeIcon.tsx
@@ -1,0 +1,12 @@
+import styled from "@emotion/styled";
+import { ImHome3 } from "react-icons/im";
+
+export const HomeIcon = styled(ImHome3)`
+  font-size: 2.4rem;
+  color: #666666;
+  @media only screen and (min-width: 768px) {
+    &:hover {
+      color: #000;
+    }
+  }
+`;
