+++ src/Components/Atoms/Hr.tsx
@@ -1,0 +1,8 @@
+import styled from "@emotion/styled";
+
+export const Line = styled.hr`
+  width: 45%;
+  color: #d9ddde;
+  height: 2px;
+  background: #d9ddde;
+`;
