+++ src/Components/Atoms/LabelTag.tsx
@@ -1,0 +1,7 @@
+import styled from "@emotion/styled";
+
+export const Label = styled.label`
+font-size: 14px;
+  text-align left;
+  margin-bottom: 0;
+`;
