+++ src/Components/Atoms/LinkedLogo.tsx
@@ -1,0 +1,9 @@
+import styled from "@emotion/styled";
+import { FaLinkedin } from "react-icons/fa";
+
+export const LinkedInIcon = styled(FaLinkedin)`
+  color: #0a66c2;
+  font-size: 26px;
+  font-weight: 500;
+  height: inherit;
+`;
