+++ src/Components/Atoms/LinkedLogoTxt.tsx
@@ -1,0 +1,9 @@
+import styled from "@emotion/styled";
+
+export const Linked = styled.h2`
+  color: #0a66c2;
+  font-size: 28px;
+  font-weight: 800;
+  display: flex;
+  justify-self: flex-start;
+`;
