+++ src/Components/Atoms/Ptag.tsx
@@ -1,0 +1,25 @@
+import styled from "@emotion/styled";
+
+export const Terms = styled.p`
+  color: #6a6a6a;
+  font-size: 12px;
+  text-align: center;
+`;
+
+export const Or = styled.p`
+  font-size: 16px;
+  color: #000;
+`;
+
+export const Ptag = styled.p`
+  font-size: 1rem;
+  line-height: 1.1;
+  font-weight: 350;
+  color: #000;
+  @media only screen and (min-width: 768px) {
+    font-size: 0.9rem;
+    margin-top: -15px;
+    line-height: 1.3;
+    font-weight: 400;
+  }
+`;
