+++ src/Components/Atoms/Span.tsx
@@ -1,0 +1,8 @@
+import styled from "@emotion/styled";
+
+export const Span = styled.span`
+  display: flex;
+  align-items: center;
+  justify-self: center;
+  margin-top: 0;
+`;
