+++ src/Components/Molecules/OrSec.tsx
@@ -1,0 +1,9 @@
+import styled from "@emotion/styled";
+
+export const OrSec = styled.span`
+    display: flex;
+    align-items: center;
+    max-width: 348px;
+    width: 100%;
+    margin: auto;
+`;
