+++ src/Components/Molecules/PasswordHolder.tsx
@@ -1,0 +1,13 @@
+import styled from "@emotion/styled";
+
+export const PassHolder = styled.div`
+    max-width: 348px;
+    width: 100%;
+    margin: auto;
+    background: #fff;
+    border: 1px solid #000;
+    align-items: center;
+    justify-content: space-between;
+    display: flex;
+    border-radius: 4px !important;
+`;
