+++ src/Components/Organisms/Footer.tsx
@@ -1,0 +1,12 @@
+import styled from "@emotion/styled";
+
+export const Footer = styled.footer`
+  display: flex;
+  margin-top: auto;
+  padding: 12px 80px;
+  justify-content: space-between;
+  align-items: center;
+  bottom: 0;
+  justify-self: flex-end;
+  width: 100vw;
+`;
