+++ src/Components/Organisms/MainDiv.tsx
@@ -1,0 +1,12 @@
+import styled from "@emotion/styled";
+
+export const MainDiv = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin: 0;
+  height: 100vh;
+  height: fit-content;
+  min-height: 100vh;
+`;
