+++ src/Components/Organisms/NavBar.tsx
@@ -1,0 +1,13 @@
+import styled from "@emotion/styled";
+
+export const NavBar = styled.div`
+  display: flex;
+  padding: 10px 10%;
+  justify-content: space-between;
+  align-items: center;
+  top: 0;
+  width: 100vw;
+  @media only screen and (min-width: 768px) {
+    padding: 10px 80px;
+  }
+`;
